move RenderComponent to separate category
